[BUGFIX]: Multithread Concurrency applies clamav.,#
Detail:
* Libclamav using configure with doc_build for build library didn't
* included with LLVM.

ReviewedBy:R.Chatsiri
AutomergeTo:

Resolves: #
Documentation: #
Release:

	modified:   src/filetypes/pe_file_controller.cpp
	modified:   src/memory/file_shm_handler.cpp
	modified:   src/memory/file_shm_handler.hpp
	modified:   src/ocl/utils/clutil_commandqueue.cpp
	modified:   src/test/Makefile.am
	modified:   src/test/file_policy_test
	modified:   src/test/scan_pe_file_acp_test
	modified:   src/test/scan_pe_file_acp_test.cpp
	modified:   src/test/test_suite
	new file:   src/test/thread_sync_test
	new file:   src/test/thread_sync_test.cpp
	modified:   src/threadconcurrency/cliprescan_pe_controller.cpp
	modified:   src/threadconcurrency/cliprescan_pe_controller.hpp
	modified:   src/threadconcurrency/cliprescan_pe_task.cpp
	modified:   src/threadsyncocl/thread_controller.cpp
	modified:   src/threadsyncocl/thread_sync.cpp
	deleted:    other.fil
	modified:   src/clibasewrapper (modified content)
	deleted:    src/filetypes/pe_prescan_controller.cpp
	deleted:    src/filetypes/pe_prescan_controller.hpp
	deleted:    src/test/file_policy_test.hpp
	deleted:    src/test/scan_pe_file_actire_parallel_test.hpp
	deleted:    src/test/thread_sync_test.hpp
	deleted:    src/threadconcurrency/.deps/monitor_controller.Tpo

+++ src/threadconcurrency/cliprescan_pe_controller.hpp
@@ -14,7 +14,7 @@ namespace controller
 
         public:
             bool initial_task_size(size_t thread_count,
-                    size_t timeout,
+                    int64_t timeout,
                     std::vector<MAPPED_FILE *> * mappe_file_pe_vec);
 
             bool task_start();
